+++ test/CMakeLists.txt
@@ -2,18 +2,16 @@
 add_subdirectory(gtest EXCLUDE_FROM_ALL)
 
 enable_testing()
-add_subdirectory(test_helper)
-add_subdirectory(test_runner)
-add_subdirectory(graph_test)
+add_subdirectory(api)
+add_subdirectory(binder)
+add_subdirectory(copy)
 add_subdirectory(c_api)
 add_subdirectory(common)
-add_subdirectory(main)
+add_subdirectory(graph_test)
 add_subdirectory(optimizer)
 add_subdirectory(planner)
 add_subdirectory(runner)
 add_subdirectory(storage)
+add_subdirectory(test_helper)
+add_subdirectory(test_runner)
 add_subdirectory(transaction)
-add_subdirectory(util_tests)
-add_subdirectory(copy)
-add_subdirectory(function/gds)
-add_subdirectory(binder)
