+++ test/api/arrow_test.cpp
@@ -1,5 +1,5 @@
+#include "api_test/api_test.h"
 #include "common/exception/runtime.h"
-#include "main_test_helper/main_test_helper.h"
 
 using namespace kuzu::common;
 using namespace kuzu::testing;
