+++ test/api/db_locking_test.cpp
@@ -2,7 +2,7 @@
 #include <sys/mman.h>
 #include <unistd.h>
 
-#include "main_test_helper/main_test_helper.h"
+#include "api_test/api_test.h"
 
 using namespace kuzu::testing;
 using namespace kuzu::common;
