+++ test/api/read_only_test.cpp
@@ -1,4 +1,4 @@
-#include "main_test_helper/main_test_helper.h"
+#include "api_test/api_test.h"
 
 using namespace kuzu::common;
 using namespace kuzu::testing;
