+++ test/api/system_config_test.cpp
@@ -1,6 +1,6 @@
+#include "api_test/api_test.h"
 #include "common/exception/buffer_manager.h"
 #include "common/system_config.h"
-#include "main_test_helper/main_test_helper.h"
 
 using namespace kuzu::common;
 using namespace kuzu::testing;
