+++ test/common/CMakeLists.txt
@@ -3,6 +3,7 @@ add_kuzu_test(types_test
         date_test.cpp
         interval_test.cpp
         null_mask_test.cpp
+        string_format.cpp
         string_test.cpp
         time_test.cpp
         timestamp_test.cpp)
