+++ test/common/string_format.cpp
