+++ test/copy/multi_copy_test.cpp
@@ -4,11 +4,11 @@
 #include <fstream>
 #include <random>
 
+#include "../include/test_helper/test_helper.h"
 #include "common/file_system/local_file_system.h"
 #include "common/string_format.h"
 #include "common/system_config.h"
-#include "graph_test/graph_test.h"
-#include "test_helper/test_helper.h"
+#include "graph_test/private_graph_test.h"
 
 namespace kuzu {
 namespace testing {
