+++ test/planner/cardinality_test.cpp
@@ -1,4 +1,4 @@
-#include "graph_test/graph_test.h"
+#include "graph_test/private_graph_test.h"
 #include "planner/operator/logical_plan_util.h"
 #include "test_runner/test_runner.h"
