+++ test/storage/node_update_test.cpp
@@ -1,4 +1,4 @@
-#include "graph_test/graph_test.h"
+#include "graph_test/private_graph_test.h"
 
 namespace kuzu {
 namespace testing {
