+++ test/test_helper/CMakeLists.txt
@@ -3,13 +3,11 @@ add_library(
         OBJECT
         test_helper.cpp
 )
-
 target_include_directories(
         test_helper
         PUBLIC
         ../include/
 )
-
 target_link_libraries(test_helper PUBLIC kuzu)
 
 add_library(
@@ -17,11 +15,9 @@ add_library(
         OBJECT
         test_helper.cpp
 )
-
 target_include_directories(
         api_test_helper
         PUBLIC
         ../include/
 )
-
 target_link_libraries(api_test_helper PUBLIC kuzu_shared)
