+++ test/transaction/current_time_test.cpp
@@ -1,5 +1,5 @@
 #include "common/types/date_t.h"
-#include "graph_test/graph_test.h"
+#include "graph_test/private_graph_test.h"
 
 namespace kuzu {
 namespace testing {
