--- test/util_tests/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_kuzu_test(utils_test
-        string_format.cpp)
-
